requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
numpy<2
pillow==10.2.0
opencv-python==4.9.0.80
scikit-learn==1.4.0
matplotlib==3.8.2
tensorflow==2.15.0
keras==2.15.0